refactor(api): extract shared platform helpers
